pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelspointspolygons"
version = "0.1.0"
description = "The P3 dataset: Pixels, Points and Polygons for Multimodal Building Vectorization"
authors = [
    { name = "Raphael Sulzer", email = "raphael.sulzer.1@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "albumentations",
    "colorlog",
    "descartes==1.1.0",
    "huggingface_hub",
    "hydra-core",
    "laspy[lazrs]",
    "matplotlib",
    "numpy==1.25.2", # this line is only necessary when installing with open3d, i.e. pytorch 2.2, not necessary with the newest pytorch
    "open3d==0.18.0",
    "pandas",
    "pycocotools",
    "rasterio",
    "scikit-image",
    "scikit-learn",
    "shapely",
    "tensorboard",
    "timm",
    "torchmetrics",
    "transformers==4.32",
    "wandb",
    # only for hisup
    "Cython",
    # only for FFL preprocessing
    "kornia",
    "skan",
    "jsmin",
    # ffl asm polygonization
    # "torch-scatter", has to be specified in requirements-torch-cuda for specific cuda version
]


[tool.setuptools]
packages = ["pixelspointspolygons"]

[tool.setuptools.package-data]
pixelspointspolygons = ["conf/*.yaml"]
